Patched the horrid collision glitch
The glitch that was causing players to sometimes clip through the ground has been patched! It was a nightmare to fix, and the code is a little bit messy. I will fix up the code later using a function, but until then, you might want to avoid the obj_player step event lol.

Changelog:
-------------------------------------------
Bug Fixes:
-Fixed a collision glitch that caused players to sometimes clip through walls if they were standing in a wall without collision.