Created moveable boxes
Changelog:
-------------------------------------------
New Features:
-Added moveable boxes called crates that can be pushed around and stood on, and have gravity and collision. When the player is pushing them, they will move slower.
-The key "R" has been mapped to resetting the room.

Known Bugs:
If the player is underneath a crate as it falls, the crate will clip through the player. This will not softlock them but it's still probably a good idea to patch this glitch. (Note that if the crate starts on the ground, there should be no way for the player to get underneath it as it is falling.)